reanme: NotificationRecord -> NotificationPayload로 이름 변경
